+++ packages/pluggableWidgets/datagrid-text-filter-web/src/DatagridTextFilter.editorConfig.ts
@@ -1,18 +1,27 @@
 import {
     ContainerProps,
     containsIcon,
+    containsIconDark,
     endsWithIcon,
+    endsWithIconDark,
     equalsIcon,
+    equalsIconDark,
     greaterThanEqualIcon,
+    greaterThanEqualIconDark,
     greaterThanIcon,
+    greaterThanIconDark,
     hidePropertiesIn,
     hidePropertyIn,
     ImageProps,
     notEqualIcon,
+    notEqualIconDark,
     Properties,
     smallerThanEqualIcon,
+    smallerThanEqualIconDark,
     smallerThanIcon,
+    smallerThanIconDark,
     startsWithIcon,
+    startsWithIconDark,
     StructurePreviewProps,
     TextProps
 } from "@mendix/piw-utils-internal";
@@ -36,7 +45,7 @@ export function getProperties(
     return defaultProperties;
 }
 
-export const getPreview = (values: DatagridTextFilterPreviewProps): StructurePreviewProps => {
+export const getPreview = (values: DatagridTextFilterPreviewProps, isDarkMode: boolean): StructurePreviewProps => {
     const adjustableByUserContainer = values.adjustable
         ? [
               {
@@ -46,7 +55,7 @@ export const getPreview = (values: DatagridTextFilterPreviewProps): StructurePre
                   children: [
                       {
                           type: "Image",
-                          document: getSvgContent(values.defaultFilter)
+                          document: getSvgContent(values.defaultFilter, isDarkMode)
                       } as ImageProps
                   ]
               } as ContainerProps,
@@ -68,7 +77,7 @@ export const getPreview = (values: DatagridTextFilterPreviewProps): StructurePre
             {
                 type: "RowLayout",
                 columnSize: "grow",
-                backgroundColor: "#FFFFFF",
+                backgroundColor: isDarkMode ? "#313131" : "#FFFFFF",
                 children: [
                     ...adjustableByUserContainer,
                     {
@@ -77,7 +86,13 @@ export const getPreview = (values: DatagridTextFilterPreviewProps): StructurePre
                         children: [
                             {
                                 type: "Text",
-                                fontColor: values.placeholder ? "#BBBBBB" : "#FFF",
+                                fontColor: values.placeholder
+                                    ? isDarkMode
+                                        ? "#A4A4A4"
+                                        : "#BBBBBB"
+                                    : isDarkMode
+                                    ? "#313131"
+                                    : "#FFF",
                                 italic: true,
                                 content: values.placeholder ? values.placeholder : "Sample"
                             } as TextProps
@@ -90,25 +105,25 @@ export const getPreview = (values: DatagridTextFilterPreviewProps): StructurePre
     };
 };
 
-function getSvgContent(type: DefaultFilterEnum): string {
+function getSvgContent(type: DefaultFilterEnum, isDarkMode: boolean): string {
     switch (type) {
         case "contains":
-            return containsIcon;
+            return isDarkMode ? containsIconDark : containsIcon;
         case "endsWith":
-            return endsWithIcon;
+            return isDarkMode ? endsWithIconDark : endsWithIcon;
         case "equal":
-            return equalsIcon;
+            return isDarkMode ? equalsIconDark : equalsIcon;
         case "notEqual":
-            return notEqualIcon;
+            return isDarkMode ? notEqualIconDark : notEqualIcon;
         case "greater":
-            return greaterThanIcon;
+            return isDarkMode ? greaterThanIconDark : greaterThanIcon;
         case "greaterEqual":
-            return greaterThanEqualIcon;
+            return isDarkMode ? greaterThanEqualIconDark : greaterThanEqualIcon;
         case "smaller":
-            return smallerThanIcon;
+            return isDarkMode ? smallerThanIconDark : smallerThanIcon;
         case "smallerEqual":
-            return smallerThanEqualIcon;
+            return isDarkMode ? smallerThanEqualIconDark : smallerThanEqualIcon;
         case "startsWith":
-            return startsWithIcon;
+            return isDarkMode ? startsWithIconDark : startsWithIcon;
     }
 }