+++ packages/pluggableWidgets/fieldset-web/src/Fieldset.editorConfig.ts
@@ -1,7 +1,7 @@
 import { DropZoneProps, RowLayoutProps, StructurePreviewProps, TextProps } from "@mendix/piw-utils-internal/dist";
 import { FieldsetPreviewProps } from "../typings/FieldsetProps";
 
-export function getPreview(values: FieldsetPreviewProps): StructurePreviewProps | null {
+export function getPreview(values: FieldsetPreviewProps, isDarkMode: boolean): StructurePreviewProps | null {
     return {
         type: "Container",
         borders: true,
@@ -9,19 +9,25 @@ export function getPreview(values: FieldsetPreviewProps): StructurePreviewProps 
         children: [
             {
                 type: "Container",
-                padding: 10,
+                backgroundColor: isDarkMode ? "#454545" : undefined,
                 children: [
                     {
-                        type: "Text",
-                        content: values.legend || "Legend",
-                        fontSize: 14,
-                        bold: true
-                    } as TextProps
+                        type: "Container",
+                        padding: 10,
+                        children: [
+                            {
+                                type: "Text",
+                                content: values.legend || "Legend",
+                                fontSize: 14,
+                                bold: true
+                            } as TextProps
+                        ]
+                    }
                 ]
             },
             {
                 type: "RowLayout",
-                backgroundColor: values.content.widgetCount > 0 ? undefined : "#F8F8F8",
+                backgroundColor: values.content.widgetCount > 0 ? undefined : isDarkMode ? undefined : "#F8F8F8",
                 children: [
                     {
                         type: "DropZone",