+++ packages/pluggableWidgets/progress-bar-web/src/ProgressBar.editorConfig.ts
@@ -7,6 +7,7 @@ import {
 } from "@mendix/piw-utils-internal";
 import { ProgressBarPreviewProps, TypeEnum } from "../typings/ProgressBarProps";
 import StructurePreviewSvg from "./assets/structure-preview.svg";
+import StructurePreviewSvgDark from "./assets/structure-preview-dark.svg";
 
 type PreviewPropsKey = keyof ProgressBarPreviewProps;
 const propKeys: { [Type in TypeEnum]: PreviewPropsKey[] } = {
@@ -103,13 +104,15 @@ export function check(values: ProgressBarPreviewProps): Problem[] {
     return errors;
 }
 
-export function getPreview(values: ProgressBarPreviewProps): StructurePreviewProps | null {
+export function getPreview(values: ProgressBarPreviewProps, isDarkMode: boolean): StructurePreviewProps | null {
     if (values.labelType === "custom") {
         return null;
     }
     return {
         type: "Image",
-        document: decodeURIComponent(StructurePreviewSvg.replace("data:image/svg+xml,", "")),
+        document: decodeURIComponent(
+            (isDarkMode ? StructurePreviewSvgDark : StructurePreviewSvg).replace("data:image/svg+xml,", "")
+        ),
         height: 16,
         width: 300
     };