+++ packages/pluggableWidgets/accessibility-helper-web/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- We added dark mode to Structure mode preview.
+
 ## [2.0.0] - 2021-09-28
 
 ### Added