+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+- We added dark mode to Structure mode preview.
+
 ## [2.0.2] - 2021-10-13
 
 ### Added